+++ lib/src/not_shared_map.dart
@@ -20,6 +20,7 @@ class NotSharedStore implements SharedStore {
   SharedMap<K, V>? getSharedMap<K, V>(
     String id, {
     SharedMapEventCallback? onInitialize,
+    SharedMapKeyCallback<K, V>? onAbsent,
     SharedMapEntryCallback<K, V>? onPut,
     SharedMapEntryCallback<K, V>? onRemove,
   }) {
@@ -29,7 +30,10 @@ class NotSharedStore implements SharedStore {
         : NotSharedMap<K, V>();
 
     o.setCallbacksDynamic<K, V>(
-        onInitialize: onInitialize, onPut: onPut, onRemove: onRemove);
+        onInitialize: onInitialize,
+        onAbsent: onAbsent,
+        onPut: onPut,
+        onRemove: onRemove);
 
     return o;
   }
@@ -74,6 +78,9 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   SharedMapEventCallback? onInitialize;
 
   @override
+  SharedMapKeyCallback<K, V>? onAbsent;
+
+  @override
   SharedMapEntryCallback<K, V>? onPut;
 
   @override
@@ -82,11 +89,16 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   @override
   void setCallbacks(
       {SharedMapEventCallback? onInitialize,
+      SharedMapKeyCallback<K, V>? onAbsent,
       SharedMapEntryCallback<K, V>? onPut,
       SharedMapEntryCallback<K, V>? onRemove}) {
     if (onInitialize != null && this.onInitialize == null) {
       this.onInitialize = onInitialize;
       onInitialize(this);
+    }
+
+    if (onAbsent != null) {
+      this.onAbsent ??= onAbsent;
     }
 
     if (onPut != null) {
@@ -101,11 +113,16 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   @override
   void setCallbacksDynamic<K1, V1>(
       {SharedMapEventCallback? onInitialize,
+      SharedMapKeyCallback<K1, V1>? onAbsent,
       SharedMapEntryCallback<K1, V1>? onPut,
       SharedMapEntryCallback<K1, V1>? onRemove}) {
     if (onInitialize != null && this.onInitialize == null) {
       this.onInitialize = onInitialize;
       onInitialize(this);
+    }
+
+    if (onAbsent is SharedMapKeyCallback<K, V>) {
+      this.onAbsent ??= onAbsent as SharedMapKeyCallback<K, V>;
     }
 
     if (onPut is SharedMapEntryCallback<K, V>) {
@@ -123,7 +140,21 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   SharedMapCached<K, V> cached({Duration? timeout}) => _cached;
 
   @override
-  V? get(K key) => _entries[key];
+  V? get(K key) {
+    var value = _entries[key];
+
+    if (value == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        value = onAbsent(key);
+        if (value != null) {
+          _entries[key] = value;
+        }
+      }
+    }
+
+    return value;
+  }
 
   @override
   List<K> keys() => _entries.keys.toList();
@@ -158,6 +189,17 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   @override
   V? putIfAbsent(K key, V? absentValue) {
     var prev = _entries[key];
+
+    if (prev == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        prev = onAbsent(key);
+        if (prev != null) {
+          _entries[key] = prev;
+        }
+      }
+    }
+
     if (prev == null) {
       if (absentValue == null) {
         return null;
@@ -176,6 +218,17 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   @override
   V? update(K key, SharedMapUpdater<K, V> updater) {
     var prev = _entries[key];
+
+    if (prev == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        prev = onAbsent(key);
+        if (prev != null) {
+          _entries[key] = prev;
+        }
+      }
+    }
+
     var value = updater(key, prev);
     return put(key, value);
   }
@@ -183,6 +236,14 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   @override
   V? remove(K key) {
     var v = _entries.remove(key);
+
+    if (v == null) {
+      var onAbsent = this.onAbsent;
+      if (onAbsent != null) {
+        v = onAbsent(key);
+      }
+    }
+
     if (v != null) {
       onRemove.callback(key, v);
     }
@@ -190,10 +251,7 @@ class NotSharedMap<K, V> implements SharedMapSync<K, V> {
   }
 
   @override
-  List<V?> removeAll(List<K> keys) {
-    var values = keys.map((k) => _entries.remove(k)).toList();
-    return values;
-  }
+  List<V?> removeAll(List<K> keys) => keys.map(remove).toList();
 
   @override
   int clear() {
@@ -255,6 +313,13 @@ class _NotSharedMapCache<K, V> implements SharedMapCached<K, V> {
       _sharedMap.onInitialize = callback;
 
   @override
+  SharedMapKeyCallback<K, V>? get onAbsent => _sharedMap.onAbsent;
+
+  @override
+  set onAbsent(SharedMapKeyCallback<K, V>? callback) =>
+      _sharedMap.onAbsent = callback;
+
+  @override
   SharedMapEntryCallback<K, V>? get onPut => _sharedMap.onPut;
 
   @override
@@ -271,18 +336,26 @@ class _NotSharedMapCache<K, V> implements SharedMapCached<K, V> {
   @override
   void setCallbacks(
           {SharedMapEventCallback? onInitialize,
+          SharedMapKeyCallback<K, V>? onAbsent,
           SharedMapEntryCallback<K, V>? onPut,
           SharedMapEntryCallback<K, V>? onRemove}) =>
       _sharedMap.setCallbacks(
-          onInitialize: onInitialize, onPut: onPut, onRemove: onRemove);
+          onInitialize: onInitialize,
+          onAbsent: onAbsent,
+          onPut: onPut,
+          onRemove: onRemove);
 
   @override
   void setCallbacksDynamic<K1, V1>(
           {SharedMapEventCallback? onInitialize,
+          SharedMapKeyCallback<K1, V1>? onAbsent,
           SharedMapEntryCallback<K1, V1>? onPut,
           SharedMapEntryCallback<K1, V1>? onRemove}) =>
       _sharedMap.setCallbacksDynamic(
-          onInitialize: onInitialize, onPut: onPut, onRemove: onRemove);
+          onInitialize: onInitialize,
+          onAbsent: onAbsent,
+          onPut: onPut,
+          onRemove: onRemove);
 
   @override
   String get id => _sharedMap.id;